refactor(svelte2tsx): snippet / slot / events / doc / svelte:options を nodes/ へ
- `nodes/snippet_hoisting.rs`: トップレベル `{#snippet}` の移送と巻き上げ可否判定
- `nodes/slot.rs`: `<slot>` の探索と component export の `slots` リテラル生成
- `nodes/component_events.rs`: `events` リテラル生成
- `nodes/component_documentation.rs`: `<!-- @component … -->` の JSDoc 化
- `nodes/svelte_options.rs`: `ast.options` 専用の emitter

`fragment_has_slot_element` / `node_has_slot_element` は runes 検出群の間にあったが
slot 探索なので `nodes/slot.rs` に置いた。

純粋な移動でロジック変更はなく、出力はバイト単位で不変。
