+++ crates/rsvelte_core/src/svelte2tsx/nodes/mod.rs
@@ -1,6 +1,11 @@
 //! Per-node analyses and emitters, mirroring `svelte2tsx/src/svelte2tsx/nodes/`.
 
+pub mod component_documentation;
+pub mod component_events;
 pub mod component_name;
 pub mod generics;
 pub mod runes_detection;
 pub mod scripts;
+pub mod slot;
+pub mod snippet_hoisting;
+pub mod svelte_options;
