+++ crates/rsvelte_core/src/svelte2tsx/svelte2tsx.rs
@@ -10,6 +10,8 @@ use crate::compiler::phases::phase1_parse::{self, ParseOptions};
 
 use super::helpers::rewrite_external_imports::rewrite_external_specifiers_in_text;
 use super::magic_string::{GenerateMapOptions, MagicString};
+use super::nodes::component_documentation::extract_component_documentation;
+use super::nodes::component_events::build_events_str;
 use super::nodes::component_name::derive_component_name;
 use super::nodes::generics::{
     compact_generic_params, extract_generics_from_script_tag, split_generic_param_names,
@@ -21,10 +23,15 @@ use super::nodes::runes_detection::{
 use super::nodes::scripts::{
     detect_top_level_await, find_instance_imports, find_script_close_tag_start,
 };
+use super::nodes::slot::{
+    build_slots_str, collect_slot_names_from_ast, escape_js_single_quoted,
+    fragment_has_slot_element,
+};
+use super::nodes::snippet_hoisting::hoist_top_level_snippets;
+use super::nodes::svelte_options::emit_svelte_options_element;
 use super::script::{ComponentEvents, ExportedNames};
 use super::template;
 use super::utils::htmlxparser::{blank_style_content, blank_style_tags, remove_orphan_scripts};
-use super::utils::lexical::{lexical_identifiers, lexical_identifiers_in_expressions};
 use super::validation::{validate_debug_tag_arguments, validate_meta_element_placement};
 
 pub use super::interfaces::{
@@ -1729,215 +1736,6 @@ pub fn svelte2tsx(
     })
 }
 
-/// Emit the `<svelte:options>` tag as a `svelteHTML.createElement(...)` call.
-/// The parser stores svelte:options in `ast.options` (not in `fragment.nodes`),
-/// so it is handled separately.
-fn emit_svelte_options_element(ast: &Root, source: &str, str: &mut MagicString) {
-    let Some(options_node) = ast.options.as_ref() else {
-        return;
-    };
-    if options_node.start >= options_node.end {
-        return;
-    }
-    // Build attribute string from options attributes
-    let mut attrs_parts = Vec::new();
-    let mut has_expression_attr = false;
-    for node in &options_node.attributes {
-        match &node.value {
-            crate::ast::template::AttributeValue::True(_) => {
-                attrs_parts.push(format!("\"{}\":true,", node.name));
-            }
-            crate::ast::template::AttributeValue::Expression(expr) => {
-                has_expression_attr = true;
-                let expr_text = slice_src(
-                    source,
-                    expr.expression.start().unwrap_or(0) as usize,
-                    expr.expression.end().unwrap_or(0) as usize,
-                );
-                attrs_parts.push(format!("\"{}\":{},", node.name, expr_text));
-            }
-            // String / mixed attribute, e.g. `<svelte:options customElement="my-el">`
-            // or `namespace="svg"`. Mirror the element-attribute Sequence path
-            // (template/mod.rs::format_attribute_node_segments): a lone expression
-            // stays a bare expression, everything else becomes a template literal.
-            // Reference: language-tools .../htmlxtojsx_v2/nodes/Attribute.ts.
-            crate::ast::template::AttributeValue::Sequence(parts) => {
-                use crate::ast::template::AttributeValuePart;
-                if parts.len() == 1
-                    && let AttributeValuePart::ExpressionTag(expr) = &parts[0]
-                {
-                    has_expression_attr = true;
-                    let expr_text = slice_src(
-                        source,
-                        expr.expression.start().unwrap_or(0) as usize,
-                        expr.expression.end().unwrap_or(0) as usize,
-                    );
-                    attrs_parts.push(format!("\"{}\":{},", node.name, expr_text));
-                } else {
-                    let mut value = String::from("`");
-                    for part in parts {
-                        match part {
-                            AttributeValuePart::Text(text) => {
-                                value.push_str(
-                                    &text
-                                        .raw
-                                        .replace('\\', "\\\\")
-                                        .replace('`', "\\`")
-                                        .replace('$', "\\$"),
-                                );
-                            }
-                            AttributeValuePart::ExpressionTag(expr) => {
-                                has_expression_attr = true;
-                                if let (Some(s), Some(e)) =
-                                    (expr.expression.start(), expr.expression.end())
-                                {
-                                    value.push_str("${");
-                                    value.push_str(slice_src(source, s as usize, e as usize));
-                                    value.push('}');
-                                }
-                            }
-                        }
-                    }
-                    value.push('`');
-                    attrs_parts.push(format!("\"{}\":{},", node.name, value));
-                }
-            }
-        }
-    }
-    let attrs_str = if attrs_parts.is_empty() {
-        String::new()
-    } else if has_expression_attr {
-        // Expression attributes: preserve source spacing
-        let extra_spaces =
-            count_tag_to_attr_spaces_in_source("svelte:options", options_node.start, source);
-        format!("{}{}", " ".repeat(extra_spaces + 1), attrs_parts.join(""))
-    } else {
-        // Bare boolean attributes only: no extra spacing
-        attrs_parts.join("")
-    };
-    let replacement = format!(
-        " {{ svelteHTML.createElement(\"svelte:options\", {{{}}});}}",
-        attrs_str
-    );
-    str.overwrite(options_node.start, options_node.end, &replacement);
-}
-
-/// Analyze and relocate top-level `{#snippet}` blocks. Non-hoistable snippets
-/// (those closing over instance-script values, or referencing a non-hoistable
-/// snippet) are moved to the top of the instance script; the returned ranges are
-/// the module-hoistable snippets, which the caller relocates to module scope.
-fn hoist_top_level_snippets(
-    ast: &Root,
-    source: &str,
-    exported_names: &ExportedNames,
-    str: &mut MagicString,
-) -> Vec<(u32, u32)> {
-    let mut hoistable_snippet_ranges: Vec<(u32, u32)> = Vec::new();
-    let mut nonhoistable_snippet_ranges: Vec<(u32, u32)> = Vec::new();
-    let module_script_present = ast.module.is_some();
-
-    // Collect every top-level snippet first so we can run a fixed-point
-    // pass over their inter-dependencies (a snippet that references the
-    // name of a non-hoistable snippet is itself non-hoistable).
-    let snippets: Vec<&crate::ast::template::SnippetBlock> = ast
-        .fragment
-        .nodes
-        .iter()
-        .filter_map(|n| {
-            if let crate::ast::template::TemplateNode::SnippetBlock(s) = n {
-                if s.start < s.end {
-                    Some(s.as_ref())
-                } else {
-                    None
-                }
-            } else {
-                None
-            }
-        })
-        .collect();
-
-    let snippet_names: Vec<String> = snippets
-        .iter()
-        .filter_map(|s| {
-            let exp_s = s.expression.start()? as usize;
-            let exp_e = s.expression.end()? as usize;
-            source.get(exp_s..exp_e).map(|s| s.to_string())
-        })
-        .collect();
-    let snippet_name_set: std::collections::HashSet<String> =
-        snippet_names.iter().cloned().collect();
-
-    // Initial blocked set: snippets that directly reference an
-    // instance-script value (or a $store of one).
-    let mut blocked = vec![false; snippets.len()];
-    if module_script_present {
-        for (i, snippet) in snippets.iter().enumerate() {
-            if !is_snippet_module_hoistable(snippet, source, exported_names) {
-                blocked[i] = true;
-            }
-        }
-
-        // Fixed-point: a snippet that references the name of a blocked
-        // snippet is itself blocked. Matches the JS reference's `while`
-        // loop in `analyzeSnippets` that grows `disallowed_values`.
-        let mut changed = true;
-        while changed {
-            changed = false;
-            for i in 0..snippets.len() {
-                if blocked[i] {
-                    continue;
-                }
-                let body_start = snippets[i].start as usize;
-                let body_end = snippets[i].end as usize;
-                if body_start >= source.len() || body_end > source.len() {
-                    continue;
-                }
-                for ident in lexical_identifiers(&source[body_start..body_end]) {
-                    if ident == snippet_names[i] {
-                        continue; // self-reference
-                    }
-                    if snippet_name_set.contains(&ident) {
-                        for (j, name) in snippet_names.iter().enumerate() {
-                            if name == &ident && blocked[j] {
-                                blocked[i] = true;
-                                changed = true;
-                                break;
-                            }
-                        }
-                        if blocked[i] {
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-    } else {
-        // No module script => everything stays inside $$render (or stays
-        // put if no instance script exists either).
-        for b in blocked.iter_mut() {
-            *b = true;
-        }
-    }
-
-    for (i, snippet) in snippets.iter().enumerate() {
-        if blocked[i] {
-            nonhoistable_snippet_ranges.push((snippet.start, snippet.end));
-        } else {
-            hoistable_snippet_ranges.push((snippet.start, snippet.end));
-        }
-    }
-
-    // Inside-target moves require an instance script to anchor against.
-    if let Some(instance) = ast.instance.as_ref() {
-        let inside_target = instance.content_offset;
-        for (s, e) in nonhoistable_snippet_ranges.iter() {
-            str.move_range(*s, *e, inside_target);
-        }
-    }
-
-    hoistable_snippet_ranges
-}
-
 /// Build the `$$props`/`$$restProps`/`$$slots` declaration text injected into
 /// the `$$render()` header for a component that references those legacy magic
 /// variables.
@@ -1968,101 +1766,6 @@ fn build_dollar_declarations(
         );
     }
     dollar_decls
-}
-
-/// Build the `slots` object literal for the component export from template info.
-fn build_slots_str(template_info: &template::TemplateInfo) -> String {
-    if template_info.slots.is_empty() {
-        "{}".to_string()
-    } else {
-        let mut slot_parts = Vec::new();
-        for (name, props) in &template_info.slots {
-            let escaped_name = escape_js_single_quoted(name);
-            if props.is_empty() {
-                slot_parts.push(format!("'{}': {{}}", escaped_name));
-            } else {
-                // Slot prop keys (the `props` strings) may also carry hyphens /
-                // spaces / quotes when they come from arbitrary `slot="…"`
-                // attributes; keep them verbatim for now since they're produced
-                // upstream from validated bindings and don't reach this site
-                // with adversarial input in practice. (issue #455, H-092)
-                slot_parts.push(format!("'{}': {{{}}}", escaped_name, props.join(", ")));
-            }
-        }
-        format!("{{{}}}", slot_parts.join(", "))
-    }
-}
-
-/// Build the `events` object literal for the component export from template info
-/// and component events.
-fn build_events_str(
-    exported_names: &ExportedNames,
-    template_info: &template::TemplateInfo,
-    events: &ComponentEvents,
-) -> String {
-    if exported_names.has_events_type {
-        "{} as unknown as $$Events".to_string()
-    } else {
-        let mut event_parts = Vec::new();
-        // Official `toDefString` order: typed-dispatcher event typings FIRST,
-        // then bubbled/forwarded events, then untyped-dispatch customEvents.
-        // Add generic event typing from createEventDispatcher<Type>() first.
-        if let Some(ref generic_type) = events.dispatcher_generic_type {
-            event_parts.push(format!(
-                "...__sveltets_2_toEventTypings<{}>()",
-                generic_type
-            ));
-        }
-        // Add element events (forwarded), reducing them exactly like the
-        // official `EventHandler` bubbled-events `Map` (event-handler.ts):
-        //   * an `Element` forward does a plain `set` (OVERWRITE) — collapsing
-        //     duplicate element forwards and clobbering any earlier component
-        //     union for that name;
-        //   * a `Component` forward CONCATS into the existing entry (so each
-        //     forwarding component instance contributes a `unionType` member).
-        // Key insertion order (first occurrence) is preserved. A single value is
-        // emitted plain; multiple values become `__sveltets_2_unionType(...)`.
-        let mut grouped: Vec<(String, Vec<String>)> = Vec::new();
-        for (name, value, kind) in &template_info.element_events {
-            match kind {
-                crate::svelte2tsx::template::ForwardedEventKind::Element => {
-                    if let Some(entry) = grouped.iter_mut().find(|(n, _)| n == name) {
-                        // Plain overwrite (official `set`): a single value.
-                        entry.1 = vec![value.clone()];
-                    } else {
-                        grouped.push((name.clone(), vec![value.clone()]));
-                    }
-                }
-                crate::svelte2tsx::template::ForwardedEventKind::Component => {
-                    if let Some(entry) = grouped.iter_mut().find(|(n, _)| n == name) {
-                        entry.1.push(value.clone());
-                    } else {
-                        grouped.push((name.clone(), vec![value.clone()]));
-                    }
-                }
-            }
-        }
-        for (name, values) in &grouped {
-            if values.len() == 1 {
-                event_parts.push(format!("'{}':{}", name, values[0]));
-            } else {
-                event_parts.push(format!(
-                    "'{}':__sveltets_2_unionType({})",
-                    name,
-                    values.join(", ")
-                ));
-            }
-        }
-        // Add custom events from dispatchers (detected during script processing)
-        for (name, value) in events.get_event_entries() {
-            event_parts.push(format!("'{}': {}", name, value));
-        }
-        if event_parts.is_empty() {
-            "{}".to_string()
-        } else {
-            format!("{{{}}}", event_parts.join(", "))
-        }
-    }
 }
 
 /// Emit the `__sveltets_Render<T>` + `$$IsomorphicComponent` component export
@@ -2167,389 +1870,4 @@ fn emit_runes_generics_component(
         closing,
         "/*\u{03A9}ignore_end\u{03A9}*/export default {safe_name};"
     );
-}
-
-/// Escape a string for use as the body of a single-quoted JS string literal.
-/// Used to interpolate slot names / slot prop keys into the generated TS output
-/// without producing invalid JS when a name carries `'`, `\\`, or control
-/// characters (issue #455, H-092).
-fn escape_js_single_quoted(s: &str) -> String {
-    let mut out = String::with_capacity(s.len());
-    for c in s.chars() {
-        match c {
-            '\\' => out.push_str("\\\\"),
-            '\'' => out.push_str("\\'"),
-            '\n' => out.push_str("\\n"),
-            '\r' => out.push_str("\\r"),
-            '\t' => out.push_str("\\t"),
-            c if (c as u32) < 0x20 => {
-                let _ = write!(out, "\\u{:04x}", c as u32);
-            }
-            c => out.push(c),
-        }
-    }
-    out
-}
-
-/// Collect slot names from the template AST.
-///
-/// Walks the fragment tree looking for `<slot>` elements and collects their names.
-/// A slot without a `name` attribute is the "default" slot.
-fn collect_slot_names_from_ast(fragment: &crate::ast::template::Fragment) -> Vec<String> {
-    let mut names = Vec::new();
-    collect_slot_names_recursive(&fragment.nodes, &mut names);
-    // Deduplicate while preserving order
-    let mut seen = std::collections::HashSet::new();
-    names.retain(|n| seen.insert(n.clone()));
-    names
-}
-
-fn collect_slot_names_recursive(
-    nodes: &[crate::ast::template::TemplateNode],
-    names: &mut Vec<String>,
-) {
-    use crate::ast::template::TemplateNode;
-    for node in nodes {
-        match node {
-            TemplateNode::SlotElement(el) => {
-                // Get slot name from the `name` attribute
-                let mut slot_name = "default".to_string();
-                for attr in &el.attributes {
-                    if let crate::ast::template::Attribute::Attribute(node) = attr
-                        && node.name == "name"
-                        && let crate::ast::template::AttributeValue::Sequence(parts) = &node.value
-                    {
-                        for part in parts {
-                            if let crate::ast::template::AttributeValuePart::Text(text) = part {
-                                slot_name = text.raw.to_string();
-                            }
-                        }
-                    }
-                }
-                names.push(slot_name);
-                collect_slot_names_recursive(&el.fragment.nodes, names);
-            }
-            TemplateNode::RegularElement(el) => {
-                collect_slot_names_recursive(&el.fragment.nodes, names);
-            }
-            TemplateNode::Component(comp) => {
-                collect_slot_names_recursive(&comp.fragment.nodes, names);
-            }
-            TemplateNode::IfBlock(block) => {
-                collect_slot_names_recursive(&block.consequent.nodes, names);
-                if let Some(ref alt) = block.alternate {
-                    collect_slot_names_recursive(&alt.nodes, names);
-                }
-            }
-            TemplateNode::EachBlock(block) => {
-                collect_slot_names_recursive(&block.body.nodes, names);
-                if let Some(ref fallback) = block.fallback {
-                    collect_slot_names_recursive(&fallback.nodes, names);
-                }
-            }
-            TemplateNode::AwaitBlock(block) => {
-                if let Some(ref pending) = block.pending {
-                    collect_slot_names_recursive(&pending.nodes, names);
-                }
-                if let Some(ref then) = block.then {
-                    collect_slot_names_recursive(&then.nodes, names);
-                }
-                if let Some(ref catch) = block.catch {
-                    collect_slot_names_recursive(&catch.nodes, names);
-                }
-            }
-            TemplateNode::KeyBlock(block) => {
-                collect_slot_names_recursive(&block.fragment.nodes, names);
-            }
-            TemplateNode::SnippetBlock(block) => {
-                collect_slot_names_recursive(&block.body.nodes, names);
-            }
-            TemplateNode::SvelteBody(el)
-            | TemplateNode::SvelteDocument(el)
-            | TemplateNode::SvelteFragment(el)
-            | TemplateNode::SvelteBoundary(el)
-            | TemplateNode::SvelteHead(el)
-            | TemplateNode::SvelteOptions(el)
-            | TemplateNode::SvelteSelf(el)
-            | TemplateNode::SvelteWindow(el) => {
-                collect_slot_names_recursive(&el.fragment.nodes, names);
-            }
-            TemplateNode::TitleElement(el) => {
-                collect_slot_names_recursive(&el.fragment.nodes, names);
-            }
-            TemplateNode::SvelteComponent(comp) => {
-                collect_slot_names_recursive(&comp.fragment.nodes, names);
-            }
-            TemplateNode::SvelteElement(el) => {
-                collect_slot_names_recursive(&el.fragment.nodes, names);
-            }
-            _ => {}
-        }
-    }
-}
-
-/// Extract `@component` documentation from HTML comments in the template.
-///
-/// Looks for comments like `<!-- @component This is documentation -->` and
-/// converts them to JSDoc format: `/** This is documentation */`.
-///
-/// Also handles multiline comments:
-/// ```html
-/// <!--
-///   @component
-///   Multi-line documentation
-/// -->
-/// ```
-fn extract_component_documentation(fragment: &crate::ast::template::Fragment) -> Option<String> {
-    use crate::ast::template::TemplateNode;
-
-    for node in &fragment.nodes {
-        if let TemplateNode::Comment(comment) = node {
-            let data = comment.data.as_str().trim();
-            if data.starts_with("@component") {
-                // Extract the documentation text after @component
-                let after_tag = data.strip_prefix("@component").unwrap();
-
-                // Official trims the whole doc *before* deciding single- vs
-                // multi-line (`componentDocumentation = data.replace('@component',
-                // '').trim()`, then `if (!doc.includes('\n'))`). So a comment
-                // whose only newlines surround a single line of text (e.g.
-                // `<!--@component\nText\n-->`) is emitted as a single-line
-                // `/** Text */`. Check the trimmed content for newlines.
-                // Mirror official `ComponentDocumentation`: the whole text is
-                // trimmed first (`data.replace('@component','').trim()`), then
-                // single- vs multi-line is decided on the trimmed content.
-                let content = after_tag.trim();
-                if content.is_empty() {
-                    return Some("/** */".to_string());
-                }
-
-                if content.contains('\n') {
-                    // Official applies `dedent-js` then maps each line to
-                    // ` *${line ? ` ${line}` : ''}`. dedent-js computes the
-                    // minimum indentation among lines that FOLLOW a newline and
-                    // carry at least one leading whitespace char (the regex
-                    // `\n[\t ]+`), i.e. it ignores the first line and ignores
-                    // zero-indent lines, then strips exactly that many leading
-                    // whitespace chars from each subsequent line that has them.
-                    let lines: Vec<&str> = content.split('\n').collect();
-                    let ws_len = |l: &str| l.len() - l.trim_start_matches([' ', '\t']).len();
-                    let size = lines[1..]
-                        .iter()
-                        .map(|l| ws_len(l))
-                        .filter(|&n| n > 0)
-                        .min()
-                        .unwrap_or(0);
-
-                    let mut result = String::from("/**\n");
-                    for (i, line) in lines.iter().enumerate() {
-                        // First line is never dedented; subsequent lines lose
-                        // exactly `size` leading whitespace chars, but only when
-                        // they actually have at least that many (regex semantics).
-                        let dedented: &str = if i > 0 && size > 0 && ws_len(line) >= size {
-                            &line[size..]
-                        } else {
-                            line
-                        };
-                        if dedented.is_empty() {
-                            result.push_str(" *\n");
-                        } else {
-                            result.push_str(" * ");
-                            result.push_str(dedented);
-                            result.push('\n');
-                        }
-                    }
-                    result.push_str(" */");
-                    return Some(result);
-                } else {
-                    return Some(format!("/** {} */", content));
-                }
-            }
-        }
-    }
-
-    None
-}
-
-/// Decide whether a top-level `{#snippet}` block is module-hoistable.
-///
-/// A snippet is module-hoistable when its body's free variables resolve only
-/// to allowed references — imports, module-script values, snippet params,
-/// or globals. References to instance-script values (`let`, `const`, etc.)
-/// block hoisting. Matches the JS reference's
-/// `hoist_to_module = (globals.size === 0 || every(isAllowedReference))`
-/// in `svelte2tsx/index.ts`.
-fn is_snippet_module_hoistable(
-    snippet: &crate::ast::template::SnippetBlock,
-    source: &str,
-    exported_names: &super::script::ExportedNames,
-) -> bool {
-    // Param names shadow outer references inside the body.
-    let mut params_set: std::collections::HashSet<String> = std::collections::HashSet::new();
-    for p in snippet.parameters.iter() {
-        if let (Some(s), Some(e)) = (p.start(), p.end()) {
-            let s = s as usize;
-            let e = e as usize;
-            if s < e && e <= source.len() {
-                for tok in lexical_identifiers(&source[s..e]) {
-                    params_set.insert(tok);
-                }
-            }
-        }
-    }
-
-    // Also exclude the snippet's own name from references (its declaration
-    // shouldn't be considered a free var of itself).
-    if let (Some(s), Some(e)) = (snippet.expression.start(), snippet.expression.end()) {
-        let s = s as usize;
-        let e = e as usize;
-        if s < e && e <= source.len() {
-            for tok in lexical_identifiers(&source[s..e]) {
-                params_set.insert(tok);
-            }
-        }
-    }
-
-    // Use the entire snippet source range. Param identifiers are excluded
-    // above; the lexical scan over the whole `{#snippet ...}` ... `{/snippet}`
-    // range is conservative but adequate for fixture cases.
-    let body_start = snippet.start;
-    let body_end = snippet.end;
-    if (body_start as usize) >= source.len() || (body_end as usize) > source.len() {
-        return true;
-    }
-    let body_text = slice_src(source, body_start as usize, body_end as usize);
-
-    // Lexical scan: any identifier in the body that resolves to an
-    // instance-script value (and isn't an import or a snippet param) blocks
-    // hoisting.
-    //
-    // We use `lexical_identifiers_in_expressions` (identifiers inside `{...}` blocks)
-    // rather than `lexical_identifiers` to avoid false positives from HTML attribute
-    // NAMES like `data-state` (where `data` or `state` follow a `-` and are not JS
-    // references).  Periscopic's scope analysis only sees JS AST nodes, not attribute
-    // name strings, so this is the correct approximation.
-    //
-    // `$name` references trigger auto-store subscription; the JS reference
-    // adds the un-prefixed `name` to `disallowed_values` via
-    // `addDisallowed(getAccessedStores())`, so any `$name` whose underlying
-    // `name` is bound in the instance script (value OR import) also blocks.
-    // Collect the snippet body's expression-context identifiers ONCE — we use
-    // `lexical_identifiers_in_expressions` (only identifiers inside `{...}`
-    // blocks) rather than the general `lexical_identifiers` to avoid false
-    // positives from HTML attribute names like `data-state` (where `state`
-    // follows a `-` and is not a JS reference). Both checks below iterate it.
-    let body_idents = lexical_identifiers_in_expressions(body_text);
-    for ident in &body_idents {
-        if params_set.contains(ident) {
-            continue;
-        }
-        if let Some(stripped) = ident.strip_prefix('$')
-            && !stripped.is_empty()
-            && !stripped.starts_with('$')
-        {
-            // Auto-store subscription targets — `addDisallowed(getAccessedStores())`
-            // in the JS reference is component-wide, so check both module
-            // and instance scopes.
-            if exported_names.instance_value_names.contains(stripped)
-                || exported_names.instance_import_names.contains(stripped)
-                || exported_names.module_value_names.contains(stripped)
-                || exported_names.module_import_names.contains(stripped)
-            {
-                return false;
-            }
-        }
-
-        if exported_names.instance_value_names.contains(ident)
-            && !exported_names.instance_import_names.contains(ident)
-        {
-            return false;
-        }
-    }
-
-    // Additional check: JS reference's `addDisallowed(implicitStoreValues.getAccessedStores())`
-    // adds the base names of ALL `$X` identifiers found in the instance script (the `is_rune`
-    // filter is broken in JS due to TypeScript parent pointers not being set, so `$props`,
-    // `$bindable`, etc. always land in `accessedStores`).  If any such base name `X` appears
-    // as a plain identifier in the snippet body's EXPRESSION context, hoisting is blocked.
-    //
-    // E.g. `{#snippet Tree}` that contains `{#snippet child({ props })}` has `props` inside
-    // a `{...}` expression block.  Meanwhile `$props()` in the instance script means `props`
-    // is in `disallowed_values` (JS) / `instance_script_loose_dollar_names` (Rust). → blocked.
-    if !exported_names.instance_script_loose_dollar_names.is_empty() {
-        for ident in &body_idents {
-            if params_set.contains(ident) {
-                continue;
-            }
-            if exported_names
-                .instance_script_loose_dollar_names
-                .contains(ident)
-            {
-                return false;
-            }
-        }
-    }
-
-    true
-}
-
-/// Count whitespace between tag name and first attribute in source.
-fn count_tag_to_attr_spaces_in_source(tag_name: &str, el_start: u32, source: &str) -> usize {
-    let name_end = el_start as usize + 1 + tag_name.len(); // +1 for '<'
-    let bytes = source.as_bytes();
-    let mut count = 0;
-    let mut i = name_end;
-    while i < source.len() {
-        let ch = bytes[i];
-        if ch == b' ' || ch == b'\t' || ch == b'\n' || ch == b'\r' {
-            count += 1;
-            i += 1;
-        } else {
-            break;
-        }
-    }
-    count
-}
-
-/// True if the template fragment contains a real `<slot>` element anywhere
-/// (recursing through elements, components, control-flow blocks, and snippets).
-/// AST-based replacement for a naive `source.contains("<slot")` scan.
-fn fragment_has_slot_element(fragment: &crate::ast::template::Fragment) -> bool {
-    fragment.nodes.iter().any(node_has_slot_element)
-}
-
-fn node_has_slot_element(node: &crate::ast::template::TemplateNode) -> bool {
-    use crate::ast::template::TemplateNode as N;
-    match node {
-        N::SlotElement(_) => true,
-        N::RegularElement(e) => fragment_has_slot_element(&e.fragment),
-        N::Component(c) => fragment_has_slot_element(&c.fragment),
-        N::SvelteComponent(c) => fragment_has_slot_element(&c.fragment),
-        N::SvelteElement(e) => fragment_has_slot_element(&e.fragment),
-        N::TitleElement(e) => fragment_has_slot_element(&e.fragment),
-        N::SvelteHead(e)
-        | N::SvelteFragment(e)
-        | N::SvelteBody(e)
-        | N::SvelteWindow(e)
-        | N::SvelteDocument(e)
-        | N::SvelteBoundary(e)
-        | N::SvelteOptions(e)
-        | N::SvelteSelf(e) => fragment_has_slot_element(&e.fragment),
-        N::IfBlock(b) => {
-            fragment_has_slot_element(&b.consequent)
-                || b.alternate.as_ref().is_some_and(fragment_has_slot_element)
-        }
-        N::EachBlock(b) => {
-            fragment_has_slot_element(&b.body)
-                || b.fallback.as_ref().is_some_and(fragment_has_slot_element)
-        }
-        N::KeyBlock(b) => fragment_has_slot_element(&b.fragment),
-        N::SnippetBlock(b) => fragment_has_slot_element(&b.body),
-        N::AwaitBlock(b) => {
-            b.pending.as_ref().is_some_and(fragment_has_slot_element)
-                || b.then.as_ref().is_some_and(fragment_has_slot_element)
-                || b.catch.as_ref().is_some_and(fragment_has_slot_element)
-        }
-        _ => false,
-    }
 }
